+++ src/pages/awss3.tsx
@@ -1,10 +1,10 @@
-import React, { ReactNode, useState, useEffect } from 'react';
+import React, { useState, useEffect } from 'react';
 import type { NextPage } from 'next';
 import Head from 'next/head';
 import { useSelector, useDispatch } from 'react-redux';
-import { v4 as uuidv4 } from 'uuid';
 import { AppState } from '../redux/store';
 import { loadNYCBridgeRatings } from '../redux/reducers/nycBridgeRatingsSlice';
+import { bridgeRatingsCsvGridColumnHeader, bridgeRatingsCsvGridRowItems } from '../utils/gridCsvHeaderRows';
 import Loading from '../components/Loading';
 
 interface NYCBridgeRatingsPageProps {
@@ -32,35 +32,6 @@ const NYCBridgeRatings: NextPage<NYCBridgeRatingsPageProps> = ({documentTitle}) 
 		setTitle(documentTitle+':\u0020NYCBridgeRatings');
 	}, [documentTitle]);
 
-	function bridgeRatingsCsvGridColumnHeader() {
-		let items:ReactNode[] = [];
-		const bd = data?.split("\n");
-		const columnHeaders = bd.shift();
-		columnHeaders.split(",").map((columnHeader: string) => (
-			items.push(
-				<div key={uuidv4()} className="table-bridge-ratings-cell table-bridge-ratings-column-header">
-					<div className="table-bridge-ratings-item">{columnHeader}</div>
-				</div>
-			)
-		))
-		return items;
-	};
-
-	function bridgeRatingsCsvGridRowItems() {
-		let items:ReactNode[] = [];
-		const bd = data?.split("\n");
-		bd.slice(1).map((column: string, index: number) => (
-			column.split(",").map((rowItem,) => (
-				items.push(
-					<div key={uuidv4()} className={`table-bridge-ratings-cell ${index % 2 === 0 ? 'bg-row-color-odd' : 'bg-row-color-even'}`}>
-						<div className="table-bridge-ratings-item">{rowItem}</div>
-					</div>
-				)
-			))
-		))
-		return items;
-	};
-
 	return (
 		<>
 			<Head>
@@ -74,7 +45,11 @@ const NYCBridgeRatings: NextPage<NYCBridgeRatingsPageProps> = ({documentTitle}) 
 
 				<h3 className="mb-3">Dataset Updated: April 1, 2022</h3>
 
-				<h4 className="mb-3">Data Source here: <a href="https://data.cityofnewyork.us/Transportation/Bridge-Ratings/4yue-vjfc" target="_blank" rel="noreferrer">https://data.cityofnewyork.us/Transportation/Bridge-Ratings/4yue-vjfc</a>.</h4>
+				<div className="word-break-all mb-3">
+					<p>
+						Data Source here: <a href="https://data.cityofnewyork.us/Transportation/Bridge-Ratings/4yue-vjfc" target="_blank" rel="noreferrer">https://data.cityofnewyork.us/Transportation/Bridge-Ratings/4yue-vjfc</a>.
+					</p>
+				</div>
 
 				<div className="mb-3">
 					<p>
@@ -110,8 +85,8 @@ const NYCBridgeRatings: NextPage<NYCBridgeRatingsPageProps> = ({documentTitle}) 
 						<div className="bg-color-ivory container-padding-border-radius-1">
 							<div className="table-bridge-ratings-wrapper">
 								<div className="table-bridge-ratings-csv-repeat-6 table-bridge-ratings-display">
-									{bridgeRatingsCsvGridColumnHeader()}
-									{bridgeRatingsCsvGridRowItems()}
+									{bridgeRatingsCsvGridColumnHeader(data)}
+									{bridgeRatingsCsvGridRowItems(data)}
 								</div>
 							</div>
 						</div>
